Promote the use of 'size_t' over 'int' and 'unsigned int', if possible. Promote also the use of explicit casts, even between size_t and (unsigned int), even if it is paranoid.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10711 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/base/AudioPluginInstance.cpp
@@ -105,7 +105,7 @@ AudioPluginInstance::toXmlString()
 
     plugin << ">" << std::endl;
 
-    for (unsigned int i = 0; i < m_ports.size(); i++)
+    for (size_t i = 0; i < m_ports.size(); i++)
     {
         plugin << "                <port id=\""
                << m_ports[i]->number
