+++ src/commands/notation/MakeChordCommand.cpp
@@ -59,13 +59,13 @@ MakeChordCommand::modifySegment()
         }
     }
 
-    for (unsigned int j = 0; j < toErase.size(); ++j) {
+    for (size_t j = 0; j < toErase.size(); ++j) {
         Segment::iterator jtr(segment.findSingle(toErase[j]));
         if (jtr != segment.end())
             segment.erase(jtr);
     }
 
-    for (unsigned int j = 0; j < toInsert.size(); ++j) {
+    for (size_t j = 0; j < toInsert.size(); ++j) {
         segment.insert(toInsert[j]);
     }
